Conflicts:
	src/main/java/com/item/finance/bean/Subject.java
	src/main/webapp/WEB-INF/yx_jsp/yx_subject_show.jsp

+++ src/main/java/com/item/finance/bean/MemberTally.java
@@ -1,12 +1,16 @@
 package com.item.finance.bean;
 
-import java.io.Serializable;
-
-import javax.persistence.*;
-
-import java.math.BigInteger;
-import java.math.BigDecimal;
 import java.util.Date;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.FetchType;
+import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.ManyToOne;
+import javax.persistence.Table;
+import javax.persistence.Temporal;
+import javax.persistence.TemporalType;
 
 
 /**
@@ -17,7 +21,7 @@ import java.util.Date;
 @Table(name="member_tally")
 public class MemberTally  {
 	private int id;
-	private BigDecimal amount;
+	private double amount;
 	private Date createDate;
 	private Date payDate;
 	private int typeId;
@@ -40,11 +44,11 @@ public class MemberTally  {
 
 
 	@Column(nullable=false, precision=10, scale=2)
-	public BigDecimal getAmount() {
+	public double getAmount() {
 		return this.amount;
 	}
 
-	public void setAmount(BigDecimal amount) {
+	public void setAmount(double amount) {
 		this.amount = amount;
 	}
